V.434l Retransmit MSG

+++ src/esp32/esp32_gps.cpp
@@ -446,7 +446,7 @@ unsigned int getGPS(void)
     if(!bGPSON)
     {
         if(meshcom_settings.node_postime > 0)
-            return meshcom_settings.node_postime;
+            return (unsigned long)meshcom_settings.node_postime;
 
         return POSINFO_INTERVAL;
     }
